+++ app/src/main/java/ru/practicum/android/diploma/di/RepositoryModule.kt
@@ -1,6 +1,7 @@
 package ru.practicum.android.diploma.di
 
 import org.koin.dsl.module
+import ru.practicum.android.diploma.data.db.converter.VacancyConverter
 import ru.practicum.android.diploma.data.search.VacanciesRepository
 import ru.practicum.android.diploma.data.search.impl.VacanciesRepositoryImpl
 import ru.practicum.android.diploma.data.vacancy.VacancyRepository
@@ -12,8 +13,11 @@ val repositoryModule = module {
         VacanciesRepositoryImpl(get())
     }
 
-    single<VacancyRepository> {
-        VacancyRepositoryImpl(get())
+    single<VacancyConverter> {
+        VacancyConverter()
     }
 
+    single<VacancyRepository> {
+        VacancyRepositoryImpl(get(), get(), get())
+    }
 }
