--- app/src/main/java/ru/practicum/android/diploma/domain/impl/ShareRepositoryImpl.kt
@@ -1,35 +1,0 @@
-package ru.practicum.android.diploma.domain.impl
-
-import ru.practicum.android.diploma.data.db.AppDatabase
-import ru.practicum.android.diploma.data.db.converter.VacancyConverter
-import ru.practicum.android.diploma.data.dto.model.favorites.ShareData
-import ru.practicum.android.diploma.domain.api.ShareRepository
-import ru.practicum.android.diploma.domain.models.Vacancy
-
-class ShareRepositoryImpl(
-    id: String,
-    private val appDatabase: AppDatabase,
-    private val vacancyConvertor: VacancyConverter
-) : ShareRepository {
-    override fun getShareData(id: String): ShareData {
-        return ShareData(
-            url = "https://api.hh.ru/vacancies/$id"
-        )
-    }
-
-    override suspend fun isFavorite(id: String): Boolean {
-        return appDatabase.favouritesVacancyDao().isFavorite(vacancyId = id)
-    }
-
-    override suspend fun insertFavouritesVacancyEntity(vacancy: Vacancy) {
-        val favouritesVacancyDao = appDatabase.favouritesVacancyDao()
-        val vacancyEntity = vacancyConvertor.mapVacancyToEntity(vacancy)
-        favouritesVacancyDao.insertFavouritesVacancyEntity(vacancyEntity)
-    }
-
-    override suspend fun deleteFavouritesVacancyEntity(vacancy: Vacancy) {
-        val favouritesVacancyDao = appDatabase.favouritesVacancyDao()
-        val vacancyEntity = vacancyConvertor.mapVacancyToEntity(vacancy)
-        favouritesVacancyDao.deleteFavouritesVacancyEntity(vacancyEntity)
-    }
-}
