+++ app/src/main/java/ru/practicum/android/diploma/ui/vacancy/VacancyFragment.kt
@@ -29,7 +29,6 @@ class VacancyFragment : Fragment() {
     private val viewModel by viewModel<VacancyViewModel>()
 
     private val binding get() = _binding!!
-    private val viewModel by viewModel<VacancyViewModel>()
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -44,17 +43,36 @@ class VacancyFragment : Fragment() {
         return binding.root
     }
 
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        viewModel.getVacancyScreenStateLiveData.observe(viewLifecycleOwner) {
-            render(it)
-        }
-
-        viewModel.getVacancyRessurces(ID_VACANCY)
-        binding.ivBack.setOnClickListener {
-            parentFragmentManager.popBackStack()
-        }
-    }
+//    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+//        super.onViewCreated(view, savedInstanceState)
+//        viewModel.getVacancyScreenStateLiveData.observe(viewLifecycleOwner) {
+//            render(it)
+//        }
+//
+//        viewModel.getVacancyRessurces(ID_VACANCY)
+//        binding.ivBack.setOnClickListener {
+//            parentFragmentManager.popBackStack()
+//        }
+//
+//        val vacancy = Vacancy("id", "mock", "mock", "mock", "mock", "mock")
+//        viewModel.observeShareState().observe(viewLifecycleOwner) { sData ->
+//            sData?.let { shareVacancy(it) }
+//        }
+//
+//        binding.ivSharing.setOnClickListener {
+//            viewModel.shareVacancy(vacancy.id)
+//        }
+//
+//        viewModel.isVacancyInFavorites(vacancy.id)
+//
+//        viewModel.isFavorite.observe(viewLifecycleOwner) { state ->
+//            updateFavoriteState(state)
+//        }
+//
+//        binding.ivFavorites.setOnClickListener {
+//            viewModel.onFavoriteClicked(vacancy)
+//        }
+//    }
 
     private fun render(state: VacancyState) {
         when (state) {
